pyproject.toml
[project]
name = "gnews-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mcp>=1.21.2",
    "httpx>=0.27.0",
    "pydantic>=2.7.0",
    "python-dotenv>=1.0.0",
    "cachetools>=5.3.3",
    "diskcache>=5.6.3",
    "fastapi>=0.121.3",
    "config>=0.5.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]
